refactor: 타이핑 훅 리팩토링
